requirements.txt
streamlit
PyMuPDF
pytesseract
sentence-transformers
torch
transformers
faiss-cpu
python-dotenv
openai
Pillow